@@ -1,6 +1,7 @@
 package org.egov.pgr.producer;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.tracer.kafka.CustomKafkaTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,19 @@ public class Producer {
     @Autowired
     private CustomKafkaTemplate<String, Object> kafkaTemplate;
 
-    public void push(String topic, Object value) {
-        kafkaTemplate.send(topic, value);
+    @Autowired
+    private MultiStateInstanceUtil multiStateInstanceUtil;
+
+    /**
+     * Publishes a message to a Kafka topic specific to the tenant.
+     *
+     * @param tenantId The unique ID of the tenant for which the message is to be published.
+     * @param topic The Kafka topic name where the message needs to be sent.
+     * @param value The message payload to be sent to the Kafka topic.
+     */
+    public void push(String tenantId, String topic, Object value) {
+        String updatedTopic = multiStateInstanceUtil.getStateSpecificTopicName(tenantId, topic);
+        log.info("The Kafka topic for the tenantId : {} is : {}", tenantId, updatedTopic);
+        this.kafkaTemplate.send(updatedTopic, value);
     }
 }
