@@ -1,6 +1,8 @@
 package org.egov.pgr.repository;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.exception.InvalidTenantIdException;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.pgr.repository.rowmapper.PGRQueryBuilder;
 import org.egov.pgr.repository.rowmapper.PGRRowMapper;
 import org.egov.pgr.util.PGRConstants;
@@ -23,26 +25,42 @@ import java.util.stream.Collectors;
 public class PGRRepository {
 
 
-    private PGRQueryBuilder queryBuilder;
+    private final PGRQueryBuilder queryBuilder;
 
-    private PGRRowMapper rowMapper;
+    private final PGRRowMapper rowMapper;
 
-    private JdbcTemplate jdbcTemplate;
+    private final JdbcTemplate jdbcTemplate;
 
+    private final MultiStateInstanceUtil multiStateInstanceUtil;
+
+    /**
+     * Constructs a PGRRepository with the required query builder, row mapper, JDBC template,
+     * and multi-state instance utility.
+     *
+     * @param queryBuilder the PGRQueryBuilder used for building database queries
+     * @param rowMapper the PGRRowMapper used for mapping rows from the database
+     * @param jdbcTemplate the JdbcTemplate used for executing SQL queries
+     * @param multiStateInstanceUtil the MultiStateInstanceUtil used for multi-state specific operations
+     */
     @Autowired
-    public PGRRepository(PGRQueryBuilder queryBuilder, PGRRowMapper rowMapper, JdbcTemplate jdbcTemplate) {
+    public PGRRepository(PGRQueryBuilder queryBuilder, PGRRowMapper rowMapper, JdbcTemplate jdbcTemplate, MultiStateInstanceUtil multiStateInstanceUtil) {
         this.queryBuilder = queryBuilder;
         this.rowMapper = rowMapper;
         this.jdbcTemplate = jdbcTemplate;
+        this.multiStateInstanceUtil = multiStateInstanceUtil;
     }
 
 
     /**
-     * searches services based on search criteria and then wraps it into serviceWrappers
-     * @param criteria
-     * @return
+     * Retrieves a list of ServiceWrapper objects based on the specified search criteria.
+     * Each ServiceWrapper contains a Service object and its associated Workflow.
+     *
+     * @param criteria the search criteria used to retrieve the services. It contains
+     *                 various filters such as tenantId, serviceCode, applicationStatus, etc.
+     * @return a list of ServiceWrapper objects containing services and their associated workflows
+     * @throws InvalidTenantIdException if the tenant ID specified in the criteria is invalid
      */
-    public List<ServiceWrapper> getServiceWrappers(RequestSearchCriteria criteria){
+    public List<ServiceWrapper> getServiceWrappers(RequestSearchCriteria criteria) throws InvalidTenantIdException {
         List<Service> services = getServices(criteria);
         List<String> serviceRequestids = services.stream().map(Service::getServiceRequestId).collect(Collectors.toList());
         Map<String, Workflow> idToWorkflowMap = new HashMap<>();
@@ -56,39 +74,58 @@ public class PGRRepository {
     }
 
     /**
-     * searches services based on search criteria
-     * @param criteria
-     * @return
+     * Retrieves a list of Service objects based on the specified search criteria.
+     *
+     * @param criteria the search criteria used to filter and retrieve the services. It may include parameters like tenantId,
+     *                 application status, locality, and other relevant filters.
+     * @return a list of Service objects that match the specified search criteria.
+     * @throws InvalidTenantIdException if the tenant ID provided in the criteria is invalid.
      */
-    public List<Service> getServices(RequestSearchCriteria criteria) {
+    public List<Service> getServices(RequestSearchCriteria criteria) throws InvalidTenantIdException {
         List<Object> preparedStmtList = new ArrayList<>();
         String query = queryBuilder.getPGRSearchQuery(criteria, preparedStmtList);
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, criteria.getTenantId());
         List<Service> services =  jdbcTemplate.query(query, preparedStmtList.toArray(), rowMapper);
         return services;
     }
 
     /**
-     * Returns the count based on the search criteria
-     * @param criteria
-     * @return
+     * Retrieves the count of records based on the specified search criteria.
+     *
+     * @param criteria the search criteria containing various filters such as tenantId,
+     *                 serviceCode, applicationStatus, locality, and more, used
+     *                 for constructing the query to count matching records.
+     * @return the total count of records that match the provided search criteria.
+     * @throws InvalidTenantIdException if the tenant ID specified in the criteria is invalid.
      */
-    public Integer getCount(RequestSearchCriteria criteria) {
+    public Integer getCount(RequestSearchCriteria criteria) throws InvalidTenantIdException {
         List<Object> preparedStmtList = new ArrayList<>();
         String query = queryBuilder.getCountQuery(criteria, preparedStmtList);
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, criteria.getTenantId());
         Integer count =  jdbcTemplate.queryForObject(query, preparedStmtList.toArray(), Integer.class);
         return count;
     }
 
-
-	public Map<String, Integer> fetchDynamicData(String tenantId) {
+    /**
+     * Fetches dynamic data for a given tenant, including the count of resolved complaints
+     * and the average resolution time.
+     *
+     * @param tenantId the tenant identifier for which dynamic data is to be fetched
+     * @return a map containing dynamic data where keys represent data types (e.g., complaints resolved,
+     *         average resolution time) and their corresponding values as integers
+     * @throws InvalidTenantIdException if the provided tenant ID is invalid
+     */
+	public Map<String, Integer> fetchDynamicData(String tenantId) throws InvalidTenantIdException {
 		List<Object> preparedStmtListCompalintsResolved = new ArrayList<>();
 		String query = queryBuilder.getResolvedComplaints(tenantId,preparedStmtListCompalintsResolved );
-
+        // Replacing schema placeholder with the schema name for the tenant id
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
 		int complaintsResolved = jdbcTemplate.queryForObject(query,preparedStmtListCompalintsResolved.toArray(),Integer.class);
 
 		List<Object> preparedStmtListAverageResolutionTime = new ArrayList<>();
 		query = queryBuilder.getAverageResolutionTime(tenantId, preparedStmtListAverageResolutionTime);
-
+        // Replacing schema placeholder with the schema name for the tenant id
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
 		int averageResolutionTime = jdbcTemplate.queryForObject(query, preparedStmtListAverageResolutionTime.toArray(),Integer.class);
 
 		Map<String, Integer> dynamicData = new HashMap<String,Integer>();
