@@ -2,6 +2,7 @@ package org.egov.pgr.service;
 
 
 import org.egov.common.contract.request.RequestInfo;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.pgr.config.PGRConfiguration;
 import org.egov.pgr.producer.Producer;
 import org.egov.pgr.repository.PGRRepository;
@@ -65,11 +66,12 @@ public class PGRService {
      * @return
      */
     public ServiceRequest create(ServiceRequest request){
+        String tenantId = request.getService().getTenantId();
         Object mdmsData = mdmsUtils.mDMSCall(request);
         validator.validateCreate(request, mdmsData);
         enrichmentService.enrichCreateRequest(request);
         workflowService.updateWorkflowStatus(request);
-        producer.push(config.getCreateTopic(),request);
+        producer.push(tenantId, config.getCreateTopic(),request);
         return request;
     }
 
@@ -80,7 +82,7 @@ public class PGRService {
      * @param criteria The search criteria containg the params on which to search
      * @return
      */
-    public List<ServiceWrapper> search(RequestInfo requestInfo, RequestSearchCriteria criteria){
+    public List<ServiceWrapper> search(RequestInfo requestInfo, RequestSearchCriteria criteria) throws InvalidTenantIdException {
         validator.validateSearch(requestInfo, criteria);
 
         enrichmentService.enrichSearchRequest(requestInfo, criteria);
@@ -128,7 +130,7 @@ public class PGRService {
         validator.validateUpdate(request, mdmsData);
         enrichmentService.enrichUpdateRequest(request);
         workflowService.updateWorkflowStatus(request);
-        producer.push(config.getUpdateTopic(),request);
+        producer.push(request.getService().getTenantId(), config.getUpdateTopic(),request);
         return request;
     }
 
@@ -138,14 +140,14 @@ public class PGRService {
      * @param criteria The search criteria containg the params for which count is required
      * @return
      */
-    public Integer count(RequestInfo requestInfo, RequestSearchCriteria criteria){
+    public Integer count(RequestInfo requestInfo, RequestSearchCriteria criteria) throws InvalidTenantIdException {
         criteria.setIsPlainSearch(false);
         Integer count = repository.getCount(criteria);
         return count;
     }
 
 
-    public List<ServiceWrapper> plainSearch(RequestInfo requestInfo, RequestSearchCriteria criteria) {
+    public List<ServiceWrapper> plainSearch(RequestInfo requestInfo, RequestSearchCriteria criteria) throws InvalidTenantIdException {
         validator.validatePlainSearch(criteria);
 
         criteria.setIsPlainSearch(true);
@@ -186,7 +188,14 @@ public class PGRService {
     }
 
 
-	public Map<String, Integer> getDynamicData(String tenantId) {
+    /**
+     * Fetches dynamic data for the specified tenant.
+     *
+     * @param tenantId The identifier of the tenant for which the dynamic data is to be retrieved.
+     * @return A map containing dynamic data, where keys are of type String and values are of type Integer.
+     * @throws InvalidTenantIdException If the provided tenantId is invalid.
+     */
+	public Map<String, Integer> getDynamicData(String tenantId) throws InvalidTenantIdException {
 		
 		Map<String,Integer> dynamicData = repository.fetchDynamicData(tenantId);
 
